// src/api/vykon/models/PerformanceRecord.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Performance Records API
 * Performance records of hospital staff (worked hours, examinations, operations, on-call shifts) for the Web-In-Cloud system
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: your_email@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 *
 * @export
 * @interface PerformanceRecord
 */
export interface PerformanceRecord {
    /**
     * Server-assigned unique identifier
     * @type {number}
     * @memberof PerformanceRecord
     */
    readonly id?: number;
    /**
     * Employee identifier (badge or HR id)
     * @type {string}
     * @memberof PerformanceRecord
     */
    employeeId?: string;
    /**
     * Full name of the employee
     * @type {string}
     * @memberof PerformanceRecord
     */
    employeeName: string;
    /**
     * Date the work was performed (ISO 8601)
     * @type {Date}
     * @memberof PerformanceRecord
     */
    date: Date;
    /**
     * Hours worked on this date
     * @type {number}
     * @memberof PerformanceRecord
     */
    hoursWorked: number;
    /**
     * Number of examinations performed
     * @type {number}
     * @memberof PerformanceRecord
     */
    examinationCount: number;
    /**
     * Number of operations performed
     * @type {number}
     * @memberof PerformanceRecord
     */
    operationCount: number;
    /**
     * Number of on-call shifts
     * @type {number}
     * @memberof PerformanceRecord
     */
    shiftCount: number;
    /**
     * Optional free-text note
     * @type {string}
     * @memberof PerformanceRecord
     */
    note?: string;
}

/**
 * Check if a given object implements the PerformanceRecord interface.
 */
export function instanceOfPerformanceRecord(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "employeeName" in value;
    isInstance = isInstance && "date" in value;
    isInstance = isInstance && "hoursWorked" in value;
    isInstance = isInstance && "examinationCount" in value;
    isInstance = isInstance && "operationCount" in value;
    isInstance = isInstance && "shiftCount" in value;

    return isInstance;
}

export function PerformanceRecordFromJSON(json: any): PerformanceRecord {
    return PerformanceRecordFromJSONTyped(json, false);
}

export function PerformanceRecordFromJSONTyped(json: any, ignoreDiscriminator: boolean): PerformanceRecord {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {

        'id': !exists(json, 'id') ? undefined : json['id'],
        'employeeId': !exists(json, 'employeeId') ? undefined : json['employeeId'],
        'employeeName': json['employeeName'],
        'date': (new Date(json['date'])),
        'hoursWorked': json['hoursWorked'],
        'examinationCount': json['examinationCount'],
        'operationCount': json['operationCount'],
        'shiftCount': json['shiftCount'],
        'note': !exists(json, 'note') ? undefined : json['note'],
    };
}

export function PerformanceRecordToJSON(value?: PerformanceRecord | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {

        'employeeId': value.employeeId,
        'employeeName': value.employeeName,
        'date': (value.date.toISOString().substr(0,10)),
        'hoursWorked': value.hoursWorked,
        'examinationCount': value.examinationCount,
        'operationCount': value.operationCount,
        'shiftCount': value.shiftCount,
        'note': value.note,
    };
}